refactor : refactoring of mailService , request validation
