+++ middlewares/RequestValidationMiddleware.js
@@ -1,0 +1,9 @@
+const { validationResult } = require("express-validator");
+
+module.exports = (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ status: "failure", errors: errors.errors });
+  }
+  next();
+};
